feat(chat): Split dependency function to ./api/functions/ module get_ai_reply
