+++ api/routes/chat.py
@@ -1,7 +1,6 @@
 from uuid import UUID
 from typing import Any
 from concurrent.futures import ThreadPoolExecutor
-import requests
 
 from fastapi import APIRouter, HTTPException
 
@@ -22,23 +21,10 @@ from crud.chat import (
     read_chats
 )
 
+from api.functions.rag import get_ai_reply
+
 router = APIRouter(prefix="/chat", tags=['Chat'])
 executor = ThreadPoolExecutor()
-
-# Pass the user query to RAG server
-# Get reply from the RAG server
-def get_ai_reply(content: str) -> str:
-    try:
-        response = requests.post(
-            f"{settings.RAG_HOST}{settings.RAG_ROUTE}",
-            json={"query": content},
-            timeout=10
-        )
-        response.raise_for_status()
-        result = response.json()
-        return result.get("reply", "No reply from RAG server")
-    except requests.RequestException as e:
-        raise HTTPException(status_code=500, detail="Exception while requesting RAG server")
 
 # Initial chat -> Create a new Chat Session
 @router.post('/', response_model=ChatResponse,
